refactored validator to make type generic
